requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
mysql-connector-python
aiosqlite

google-cloud-aiplatform[agent_engines,adk]>=1.112.0
google-adk>=0.1.0
google-genai

langchain>=0.3.0
langchain-community